Fixes to the ACL model
