+++ cognee/infrastructure/data/models/DatasetData.py
@@ -1,14 +1,19 @@
 from uuid import uuid4
 from datetime import datetime, timezone
-from sqlalchemy import Column, DateTime, UUID, ForeignKey
+from sqlalchemy import Column, DateTime, UUID, ForeignKey, PrimaryKeyConstraint, UniqueConstraint
+from sqlalchemy.orm import relationship
+
 from cognee.infrastructure.databases.relational import Base
 
 class DatasetData(Base):
     __tablename__ = "dataset_data"
 
-    id = Column(UUID, primary_key = True, default = uuid4())
+    id = Column(UUID, primary_key=True, default=uuid4)
+    created_at = Column(DateTime, default=datetime.now(timezone.utc))
+    dataset_id = Column(UUID, ForeignKey("dataset.id"), nullable=False)
+    data_id = Column(UUID, ForeignKey("data.id"), nullable=False)
+    __table_args__ = (
+        UniqueConstraint('dataset_id', 'data_id', name='uix_dataset_data'),
+    )
 
-    created_at = Column(DateTime, default = datetime.now(timezone.utc))
-
-    dataset_id = Column("dataset", UUID, ForeignKey("dataset.id"), primary_key = True)
-    data_id = Column("data", UUID, ForeignKey("data.id"), primary_key = True)
+    acls = relationship('ACL', back_populates='document')
